some config adjustments and mustache working
