+++ .env.example
@@ -23,3 +23,6 @@ FIREBASE_CLIENT_EMAIL=""
 FIREBASE_CLIENT_CERT_URL=""
 FIREBASE_DATABASE_URL=""
 FIREBASE_STORAGE_BUCKET_URL=""
+
+# http
+HTTP_PORT=3000
