+++ src/api/sample/sampleRest.js
@@ -2,15 +2,19 @@ const express = require('express')
 const sampleRest = express.Router()
 const httpResponseUtils = require('../../utils/httpResponseUtils')
 const sampleService = require('./sampleService')
-const fsUtils = require('../../utils/fsUtils')
 const firebase = require('../../utils/firebaseUtils')
 const fileUtils = require('../../utils/fileUtils')
 const loggerUtils = require('../../utils/loggerUtils')
 const smtpUtils = require('../../utils/smtpUtils')
 const httpRequestUtils = require('../../utils/httpRequestUtils')
+const mustacheUtils = require('../../utils/mustacheUtils')
 
 sampleRest.get('/mustache', (req, res) => {
-  sampleService.helloWorld().then((response) => {
+  const params = {
+    id: 30
+  }
+
+  mustacheUtils.getTemplateSQL('getAllUsers', params).then(response => {
     httpResponseUtils.json(res, response)
   })
 })
@@ -23,8 +27,9 @@ sampleRest.get('/firebase', (req, res) => {
   })
 })
 
-sampleRest.get('/firebase/upload', fileUtils.single('file'), (req, res) => {
+sampleRest.get('/firebase/upload', fileUtils.multer.single('file'), (req, res) => {
   console.log('Upload Image')
+  //
   const file = req.file
   if (file) {
     firebase.uploadFile(file).then((success) => {
@@ -52,7 +57,7 @@ sampleRest.get('/helloWorld', (req, res) => {
 })
 
 sampleRest.get('/fs', (req, res) => {
-  fsUtils.writeFile('./test.txt', 'hello!').then(response => {
+  fileUtils.writeFile('./test.txt', 'hello!').then(response => {
     console.log(response)
   })
 })
