+++ src/config/i18n-locales/pt-BR.json
@@ -1,0 +1,6 @@
+{
+  "Hello": "Ola marilene! hehe!",
+  "system_error": "Hillston, temos um problema.",
+  "route_not_found %s": "Rota '%s' não foi encontrada.",
+  "generic_error": "generic_error"
+}
