--- src/config/locales/en.json
@@ -1,1 +1,0 @@
-{}
