--- src/config/locales/pt-BR.json
@@ -1,3 +1,0 @@
-{
-	"Hello": "Hello"
-}
