--- src/utils/fsUtils.js
@@ -1,30 +1,0 @@
-const fs = require('fs')
-
-const readFile = (path, opts = 'utf8') => {
-  return new Promise((resolve, reject) => {
-    fs.readFile(path, opts, (err, data) => {
-      if (err) {
-        reject(err)
-      } else {
-        resolve(data)
-      }
-    })
-  })
-}
-
-const writeFile = (path, data, opts = 'utf8') => {
-  return new Promise((resolve, reject) => {
-    fs.writeFile(path, data, opts, (err) => {
-      if (err) {
-        reject(err)
-      } else {
-        resolve()
-      }
-    })
-  })
-}
-
-module.exports = {
-  readFile,
-  writeFile
-}
